feat: basic connections
